+++ libs/sdk-core/src/lib/db/search/search.ts
@@ -29,7 +29,12 @@ export const find = (
 
   params['shards'] = nuclia.currentShards?.[kbid] || [];
   const endpoint = `${path}/find`;
-  nuclia.events?.log('lastQuery', { endpoint, params, nucliaOptions: nuclia.options });
+  nuclia.events?.log('lastQuery', {
+    endpoint,
+    params: { ...params, ...options },
+    headers: nuclia.rest.getHeaders(useGet ? 'GET' : 'POST', path),
+    nucliaOptions: nuclia.options,
+  });
   const searchMethod = useGet
     ? nuclia.rest.get<Response>(`${endpoint}?${serialize(params, options)}`, undefined, true)
     : nuclia.rest.post<Response>(endpoint, { ...params, ...options }, undefined, true);
@@ -92,7 +97,12 @@ export const search = (
   };
   params['shards'] = nuclia.currentShards?.[kbid] || [];
 
-  nuclia.events?.log('lastQuery', { endpoint: `${path}/search`, params, nucliaOptions: nuclia.options });
+  nuclia.events?.log('lastQuery', {
+    endpoint: `${path}/search`,
+    headers: nuclia.rest.getHeaders(useGet ? 'GET' : 'POST', path),
+    params: { ...params, ...options },
+    nucliaOptions: nuclia.options,
+  });
   const searchMethod = useGet
     ? nuclia.rest.get<Search.Results | IErrorResponse>(`${path}/search?${serialize(params, options)}`)
     : nuclia.rest.post<Search.Results | IErrorResponse>(`${path}/search`, { ...params, ...options });